+++ mobile/src/components/SmartWorkspaceSourceDrawer.tsx
@@ -2,8 +2,8 @@ import { useEffect, useMemo, useRef, useState } from 'react'
 import {
   ActivityIndicator,
   FlatList,
+  InteractionManager,
   Pressable,
-  StyleSheet,
   Text,
   TextInput,
   View
@@ -25,10 +25,15 @@ import {
 } from '../tasks/smart-source-paste-intent'
 import { useSmartWorkspaceSource } from '../tasks/use-smart-workspace-source'
 import type { MobileComposerSource } from '../tasks/use-mobile-composer-source'
-import { colors, radii, spacing, typography } from '../theme/mobile-theme'
-import { BottomDrawer, BOTTOM_DRAWER_HIDE_DURATION_MS } from './BottomDrawer'
+import { colors } from '../theme/mobile-theme'
+import { BottomDrawer } from './BottomDrawer'
+import { smartWorkspaceSourceDrawerStyles as styles } from './smart-workspace-source-drawer-styles'
 import { SmartSourceModeIcon } from './SmartSourceModeIcon'
 import { SmartWorkspaceSourceRow } from './SmartWorkspaceSourceRow'
+
+// Why: match MobileSearchField — native autoFocus alone often fails to raise
+// the soft keyboard when the drawer is mid-present animation.
+const SOURCE_INPUT_FOCUS_DELAY_MS = 120
 
 type Props = {
   visible: boolean
@@ -58,6 +63,7 @@ export function SmartWorkspaceSourceDrawer({
   const availableModes = useMemo(() => resolveAvailableSmartModes(availability), [availability])
   const [mode, setMode] = useState<SmartNameMode>(() => resolveDefaultSmartMode(availability))
   const [mrStateFilter, setMrStateFilter] = useState<MrStateFilter>('opened')
+  const inputRef = useRef<TextInput>(null)
   // Why: read latest availability inside the open effect without making it a
   // reactive dep (the object is recreated each render), so re-seeding happens
   // only on open, not on every availability recompute.
@@ -68,6 +74,26 @@ export function SmartWorkspaceSourceDrawer({
   useEffect(() => {
     if (visible) {
       setMode(resolveDefaultSmartMode(availabilityRef.current))
+    }
+  }, [visible])
+
+  // Why: focus after open interactions settle so the keyboard appears and the
+  // caret lands in the docked field (same value as the form via composer.name).
+  useEffect(() => {
+    if (!visible) {
+      return
+    }
+    let timeout: ReturnType<typeof setTimeout> | undefined
+    const task = InteractionManager.runAfterInteractions(() => {
+      timeout = setTimeout(() => {
+        inputRef.current?.focus()
+      }, SOURCE_INPUT_FOCUS_DELAY_MS)
+    })
+    return () => {
+      task.cancel()
+      if (timeout) {
+        clearTimeout(timeout)
+      }
     }
   }, [visible])
 
@@ -99,10 +125,6 @@ export function SmartWorkspaceSourceDrawer({
     linearWorkspaceId,
     repos
   })
-
-  function closeSoon(): void {
-    setTimeout(onClose, BOTTOM_DRAWER_HIDE_DURATION_MS)
-  }
 
   function handleSelectRow(row: SourceRow): void {
     switch (row.kind) {
@@ -154,272 +176,141 @@ export function SmartWorkspaceSourceDrawer({
   const showEmpty =
     !loading && !error && !needsGitHubRemote && effectiveMode !== 'text' && rows.length === 0
 
+  const modeTabs = SMART_MODE_OPTIONS.filter((option: SmartModeOption) =>
+    availableModes.includes(option.id)
+  )
+
   return (
     <BottomDrawer
       visible={visible}
       onClose={onClose}
       dragContentToDismiss={false}
       contentScrollable={false}
+      fillAvailable
     >
-      <View style={styles.header}>
-        <Text style={styles.title}>Name or 'Create From'</Text>
-        <Pressable onPress={closeSoon} hitSlop={8}>
-          <Text style={styles.done}>Done</Text>
-        </Pressable>
-      </View>
-
-      <TextInput
-        style={styles.search}
-        value={composer.name}
-        onChangeText={composer.setName}
-        placeholder="Type a name or search a source"
-        placeholderTextColor={colors.textMuted}
-        autoCapitalize="none"
-        autoCorrect={false}
-        autoFocus
-      />
-
-      <View style={styles.tabRow}>
-        {SMART_MODE_OPTIONS.filter((option: SmartModeOption) =>
-          availableModes.includes(option.id)
-        ).map((option) => {
-          const selected = option.id === effectiveMode
-          const tint = selected ? colors.textPrimary : colors.textSecondary
-          return (
-            <Pressable
-              key={option.id}
-              style={[styles.tab, selected && styles.tabSelected]}
-              onPress={() => setMode(option.id)}
-            >
-              <SmartSourceModeIcon icon={option.icon} color={tint} />
-              <Text style={[styles.tabText, selected && styles.tabTextSelected]}>
-                {option.label}
-              </Text>
-            </Pressable>
-          )
-        })}
-      </View>
-
-      {effectiveMode === 'gitlab' ? (
-        <View style={styles.chipRow}>
-          {MR_STATE_FILTER_OPTIONS.map((option) => {
-            const selected = option.id === mrStateFilter
-            return (
-              <Pressable
-                key={option.id}
-                style={[styles.chip, selected && styles.chipSelected]}
-                onPress={() => setMrStateFilter(option.id)}
-              >
-                <Text style={[styles.chipText, selected && styles.chipTextSelected]}>
-                  {option.label}
-                </Text>
-              </Pressable>
-            )
-          })}
+      {/* Why: column with results flex:1 + dock flex-shrink:0 at the end.
+          Fill sheet height + marginBottom place this column on the keyboard
+          top; dock must stay a non-flex sibling so FlatList cannot clip it. */}
+      <View style={styles.root}>
+        <View style={styles.header}>
+          <Text style={styles.title}>Name or &apos;Create From&apos;</Text>
+          <Pressable onPress={onClose} hitSlop={8}>
+            <Text style={styles.done}>Done</Text>
+          </Pressable>
         </View>
-      ) : null}
 
-      {crossRepoPrompt ? (
-        <View style={styles.crossRepo}>
-          <Text style={styles.crossRepoText}>
-            This item lives in {crossRepoPrompt.link.slug.owner}/{crossRepoPrompt.link.slug.repo}.
-          </Text>
-          <View style={styles.crossRepoActions}>
-            <Pressable style={styles.crossRepoDismiss} onPress={dismissCrossRepoPrompt}>
-              <Text style={styles.crossRepoDismissText}>Cancel</Text>
-            </Pressable>
-            <Pressable style={styles.crossRepoSwitch} onPress={() => void handleAcceptCrossRepo()}>
-              <Text style={styles.crossRepoSwitchText}>
-                Switch to {crossRepoPrompt.matchingRepo.displayName}
+        <View style={styles.results}>
+          {crossRepoPrompt ? (
+            <View style={styles.crossRepo}>
+              <Text style={styles.crossRepoText}>
+                This item lives in {crossRepoPrompt.link.slug.owner}/
+                {crossRepoPrompt.link.slug.repo}.
               </Text>
-            </Pressable>
-          </View>
-        </View>
-      ) : null}
-
-      {!sshReady && effectiveMode !== 'text' && effectiveMode !== 'linear' ? (
-        <Text style={styles.notice}>Connect the repository to search sources.</Text>
-      ) : needsGitHubRemote ? (
-        <Text style={styles.notice}>
-          This SSH repo needs a GitHub remote to list issues and PRs.
-        </Text>
-      ) : error ? (
-        <Text style={styles.errorNotice}>{error}</Text>
-      ) : null}
-
-      <FlatList
-        data={rows}
-        keyExtractor={(row) => row.value}
-        style={styles.list}
-        keyboardShouldPersistTaps="handled"
-        nestedScrollEnabled
-        ListFooterComponent={
-          loading ? (
-            <View style={styles.loading}>
-              <ActivityIndicator size="small" color={colors.textSecondary} />
+              <View style={styles.crossRepoActions}>
+                <Pressable style={styles.crossRepoDismiss} onPress={dismissCrossRepoPrompt}>
+                  <Text style={styles.crossRepoDismissText}>Cancel</Text>
+                </Pressable>
+                <Pressable
+                  style={styles.crossRepoSwitch}
+                  onPress={() => void handleAcceptCrossRepo()}
+                >
+                  <Text style={styles.crossRepoSwitchText}>
+                    Switch to {crossRepoPrompt.matchingRepo.displayName}
+                  </Text>
+                </Pressable>
+              </View>
             </View>
-          ) : showEmpty ? (
-            <Text style={styles.empty}>{emptyHint || 'No results found.'}</Text>
-          ) : null
-        }
-        renderItem={({ item }) => (
-          <SmartWorkspaceSourceRow row={item} onPress={() => handleSelectRow(item)} />
-        )}
-      />
+          ) : null}
+
+          {!sshReady && effectiveMode !== 'text' && effectiveMode !== 'linear' ? (
+            <Text style={styles.notice}>Connect the repository to search sources.</Text>
+          ) : needsGitHubRemote ? (
+            <Text style={styles.notice}>
+              This SSH repo needs a GitHub remote to list issues and PRs.
+            </Text>
+          ) : error ? (
+            <Text style={styles.errorNotice}>{error}</Text>
+          ) : null}
+
+          <FlatList
+            data={rows}
+            keyExtractor={(row) => row.value}
+            style={styles.list}
+            contentContainerStyle={styles.listContent}
+            keyboardShouldPersistTaps="handled"
+            keyboardDismissMode="none"
+            nestedScrollEnabled
+            ListFooterComponent={
+              loading ? (
+                <View style={styles.loading}>
+                  <ActivityIndicator size="small" color={colors.textSecondary} />
+                </View>
+              ) : showEmpty ? (
+                <Text style={styles.empty}>{emptyHint || 'No results found.'}</Text>
+              ) : rows.length === 0 && effectiveMode === 'text' ? (
+                <Text style={styles.empty}>Type a workspace name in the field below.</Text>
+              ) : null
+            }
+            renderItem={({ item }) => (
+              <SmartWorkspaceSourceRow row={item} onPress={() => handleSelectRow(item)} />
+            )}
+          />
+        </View>
+
+        <View style={styles.dock}>
+          {effectiveMode === 'gitlab' ? (
+            <View style={styles.chipRow}>
+              {MR_STATE_FILTER_OPTIONS.map((option) => {
+                const selected = option.id === mrStateFilter
+                return (
+                  <Pressable
+                    key={option.id}
+                    style={[styles.chip, selected && styles.chipSelected]}
+                    onPress={() => setMrStateFilter(option.id)}
+                  >
+                    <Text style={[styles.chipText, selected && styles.chipTextSelected]}>
+                      {option.label}
+                    </Text>
+                  </Pressable>
+                )
+              })}
+            </View>
+          ) : null}
+
+          <View style={styles.tabRow}>
+            {modeTabs.map((option) => {
+              const selected = option.id === effectiveMode
+              const tint = selected ? colors.textPrimary : colors.textSecondary
+              return (
+                <Pressable
+                  key={option.id}
+                  style={[styles.tab, selected && styles.tabSelected]}
+                  onPress={() => setMode(option.id)}
+                >
+                  <SmartSourceModeIcon icon={option.icon} color={tint} />
+                  <Text style={[styles.tabText, selected && styles.tabTextSelected]}>
+                    {option.label}
+                  </Text>
+                </Pressable>
+              )
+            })}
+          </View>
+
+          <TextInput
+            ref={inputRef}
+            style={styles.search}
+            value={composer.name}
+            onChangeText={composer.setName}
+            placeholder="Type a name or search a source"
+            placeholderTextColor={colors.textMuted}
+            autoCapitalize="none"
+            autoCorrect={false}
+            returnKeyType="done"
+            onSubmitEditing={onClose}
+            blurOnSubmit={false}
+          />
+        </View>
+      </View>
     </BottomDrawer>
   )
 }
-
-const styles = StyleSheet.create({
-  header: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: spacing.xs,
-    paddingBottom: spacing.sm
-  },
-  title: {
-    fontSize: 15,
-    fontWeight: '600',
-    color: colors.textPrimary
-  },
-  done: {
-    fontSize: typography.bodySize,
-    fontWeight: '600',
-    color: colors.accentBlue
-  },
-  search: {
-    backgroundColor: colors.bgRaised,
-    color: colors.textPrimary,
-    borderRadius: radii.input,
-    paddingHorizontal: spacing.md,
-    paddingVertical: spacing.sm,
-    fontSize: typography.bodySize,
-    borderWidth: 1,
-    borderColor: colors.borderSubtle,
-    marginBottom: spacing.sm
-  },
-  tabRow: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    gap: spacing.xs,
-    marginBottom: spacing.sm
-  },
-  tab: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: spacing.xs,
-    paddingHorizontal: spacing.sm + 2,
-    paddingVertical: spacing.xs + 2,
-    borderRadius: radii.button,
-    borderWidth: 1,
-    borderColor: colors.borderSubtle
-  },
-  tabSelected: {
-    backgroundColor: colors.bgPanel,
-    borderColor: colors.textSecondary
-  },
-  tabText: {
-    fontSize: 13,
-    color: colors.textSecondary
-  },
-  tabTextSelected: {
-    color: colors.textPrimary,
-    fontWeight: '600'
-  },
-  chipRow: {
-    flexDirection: 'row',
-    gap: spacing.xs,
-    marginBottom: spacing.sm
-  },
-  chip: {
-    paddingHorizontal: spacing.md,
-    paddingVertical: spacing.xs,
-    borderRadius: radii.button,
-    borderWidth: 1,
-    borderColor: colors.borderSubtle
-  },
-  chipSelected: {
-    backgroundColor: colors.bgPanel,
-    borderColor: colors.textSecondary
-  },
-  chipText: {
-    fontSize: 12,
-    color: colors.textSecondary
-  },
-  chipTextSelected: {
-    color: colors.textPrimary,
-    fontWeight: '600'
-  },
-  crossRepo: {
-    backgroundColor: colors.bgRaised,
-    borderRadius: radii.input,
-    borderWidth: 1,
-    borderColor: colors.borderSubtle,
-    padding: spacing.md,
-    marginBottom: spacing.sm,
-    gap: spacing.sm
-  },
-  crossRepoText: {
-    fontSize: 13,
-    color: colors.textSecondary
-  },
-  crossRepoActions: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    gap: spacing.sm
-  },
-  crossRepoDismiss: {
-    paddingHorizontal: spacing.md,
-    paddingVertical: spacing.xs + 2,
-    borderRadius: radii.button,
-    borderWidth: 1,
-    borderColor: colors.borderSubtle
-  },
-  crossRepoDismissText: {
-    fontSize: 13,
-    color: colors.textSecondary
-  },
-  crossRepoSwitch: {
-    paddingHorizontal: spacing.md,
-    paddingVertical: spacing.xs + 2,
-    borderRadius: radii.button,
-    backgroundColor: colors.bgPanel,
-    borderWidth: 1,
-    borderColor: colors.textSecondary
-  },
-  crossRepoSwitchText: {
-    fontSize: 13,
-    fontWeight: '600',
-    color: colors.textPrimary
-  },
-  notice: {
-    fontSize: 12,
-    color: colors.textMuted,
-    paddingHorizontal: spacing.xs,
-    paddingBottom: spacing.sm
-  },
-  errorNotice: {
-    fontSize: 12,
-    color: colors.statusRed,
-    paddingHorizontal: spacing.xs,
-    paddingBottom: spacing.sm
-  },
-  list: {
-    backgroundColor: colors.bgPanel,
-    borderRadius: radii.card,
-    overflow: 'hidden',
-    maxHeight: 420,
-    flexGrow: 0
-  },
-  loading: {
-    paddingVertical: spacing.lg,
-    alignItems: 'center'
-  },
-  empty: {
-    paddingVertical: spacing.lg,
-    textAlign: 'center',
-    color: colors.textMuted,
-    fontSize: 13
-  }
-})
