+++ mobile/src/components/bottom-drawer-constants.ts
@@ -1,0 +1,1 @@
+export const BOTTOM_DRAWER_HIDE_DURATION_MS = 150
